Module aws_sdk_nimble::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AcceptEulas

Fluent builder constructing a request to AcceptEulas.

CreateLaunchProfile

Fluent builder constructing a request to CreateLaunchProfile.

CreateStreamingImage

Fluent builder constructing a request to CreateStreamingImage.

CreateStreamingSession

Fluent builder constructing a request to CreateStreamingSession.

CreateStreamingSessionStream

Fluent builder constructing a request to CreateStreamingSessionStream.

CreateStudio

Fluent builder constructing a request to CreateStudio.

CreateStudioComponent

Fluent builder constructing a request to CreateStudioComponent.

DeleteLaunchProfile

Fluent builder constructing a request to DeleteLaunchProfile.

DeleteLaunchProfileMember

Fluent builder constructing a request to DeleteLaunchProfileMember.

DeleteStreamingImage

Fluent builder constructing a request to DeleteStreamingImage.

DeleteStreamingSession

Fluent builder constructing a request to DeleteStreamingSession.

DeleteStudio

Fluent builder constructing a request to DeleteStudio.

DeleteStudioComponent

Fluent builder constructing a request to DeleteStudioComponent.

DeleteStudioMember

Fluent builder constructing a request to DeleteStudioMember.

GetEula

Fluent builder constructing a request to GetEula.

GetLaunchProfile

Fluent builder constructing a request to GetLaunchProfile.

GetLaunchProfileDetails

Fluent builder constructing a request to GetLaunchProfileDetails.

GetLaunchProfileInitialization

Fluent builder constructing a request to GetLaunchProfileInitialization.

GetLaunchProfileMember

Fluent builder constructing a request to GetLaunchProfileMember.

GetStreamingImage

Fluent builder constructing a request to GetStreamingImage.

GetStreamingSession

Fluent builder constructing a request to GetStreamingSession.

GetStreamingSessionStream

Fluent builder constructing a request to GetStreamingSessionStream.

GetStudio

Fluent builder constructing a request to GetStudio.

GetStudioComponent

Fluent builder constructing a request to GetStudioComponent.

GetStudioMember

Fluent builder constructing a request to GetStudioMember.

ListEulaAcceptances

Fluent builder constructing a request to ListEulaAcceptances.

ListEulas

Fluent builder constructing a request to ListEulas.

ListLaunchProfileMembers

Fluent builder constructing a request to ListLaunchProfileMembers.

ListLaunchProfiles

Fluent builder constructing a request to ListLaunchProfiles.

ListStreamingImages

Fluent builder constructing a request to ListStreamingImages.

ListStreamingSessions

Fluent builder constructing a request to ListStreamingSessions.

ListStudioComponents

Fluent builder constructing a request to ListStudioComponents.

ListStudioMembers

Fluent builder constructing a request to ListStudioMembers.

ListStudios

Fluent builder constructing a request to ListStudios.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

PutLaunchProfileMembers

Fluent builder constructing a request to PutLaunchProfileMembers.

PutStudioMembers

Fluent builder constructing a request to PutStudioMembers.

StartStreamingSession

Fluent builder constructing a request to StartStreamingSession.

StartStudioSSOConfigurationRepair

Fluent builder constructing a request to StartStudioSSOConfigurationRepair.

StopStreamingSession

Fluent builder constructing a request to StopStreamingSession.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateLaunchProfile

Fluent builder constructing a request to UpdateLaunchProfile.

UpdateLaunchProfileMember

Fluent builder constructing a request to UpdateLaunchProfileMember.

UpdateStreamingImage

Fluent builder constructing a request to UpdateStreamingImage.

UpdateStudio

Fluent builder constructing a request to UpdateStudio.

UpdateStudioComponent

Fluent builder constructing a request to UpdateStudioComponent.