Module aws_sdk_nimble::model

Modules

active_directory_computer_attribute

See ActiveDirectoryComputerAttribute

active_directory_configuration

See ActiveDirectoryConfiguration

compute_farm_configuration

See ComputeFarmConfiguration

eula

See Eula

eula_acceptance

See EulaAcceptance

launch_profile

See LaunchProfile

launch_profile_initialization

See LaunchProfileInitialization

launch_profile_initialization_active_directory

See LaunchProfileInitializationActiveDirectory

launch_profile_initialization_script

See LaunchProfileInitializationScript

launch_profile_membership

See LaunchProfileMembership

license_service_configuration

See LicenseServiceConfiguration

new_launch_profile_member

See NewLaunchProfileMember

new_studio_member

See NewStudioMember

script_parameter_key_value

See ScriptParameterKeyValue

shared_file_system_configuration

See SharedFileSystemConfiguration

stream_configuration

See StreamConfiguration

stream_configuration_create

See StreamConfigurationCreate

streaming_image

See StreamingImage

streaming_image_encryption_configuration

See StreamingImageEncryptionConfiguration

streaming_session

See StreamingSession

streaming_session_stream

See StreamingSessionStream

studio

See Studio

studio_component

See StudioComponent

studio_component_configuration

See StudioComponentConfiguration

studio_component_initialization_script

See StudioComponentInitializationScript

studio_component_summary

See StudioComponentSummary

studio_encryption_configuration

See StudioEncryptionConfiguration

studio_membership

See StudioMembership

Structs

ActiveDirectoryComputerAttribute

An LDAP attribute of an Active Directory computer account, in the form of a name:value pair.

ActiveDirectoryConfiguration

The configuration for a Microsoft Active Directory (Microsoft AD) studio resource.

ComputeFarmConfiguration

The configuration for a render farm that is associated with a studio resource.

Eula

Represents a EULA resource.

EulaAcceptance

LaunchProfile

LaunchProfileInitialization

LaunchProfileInitializationActiveDirectory

LaunchProfileInitializationScript

LaunchProfileMembership

LicenseServiceConfiguration

The configuration for a license service that is associated with a studio resource.

NewLaunchProfileMember

NewStudioMember

ScriptParameterKeyValue

A parameter for a studio component script, in the form of a key:value pair.

SharedFileSystemConfiguration

The configuration for a shared file storage system that is associated with a studio resource.

StreamConfiguration

A configuration for a streaming session.

StreamConfigurationCreate

StreamingImage

StreamingImageEncryptionConfiguration

StreamingSession

StreamingSessionStream

Studio

StudioComponent

A network that is used by a studio’s users and workflows, including render farm, Active Directory, licensing, and file system.

StudioComponentConfiguration

The configuration of the studio component, based on component type.

StudioComponentInitializationScript

Initialization scripts for studio components.

StudioComponentSummary

StudioEncryptionConfiguration

Configuration of the encryption method that is used for the studio.

StudioMembership

Enums

LaunchProfilePersona

LaunchProfilePlatform

LaunchProfileState

LaunchProfileStatusCode

StreamingClipboardMode

StreamingImageEncryptionConfigurationKeyType

StreamingImageState

StreamingImageStatusCode

The status code.

StreamingInstanceType

StreamingSessionState

The streaming session state.

StreamingSessionStatusCode

StreamingSessionStreamState

StreamingSessionStreamStatusCode

StudioComponentInitializationScriptRunContext

StudioComponentState

The current state of the studio component resource.

While a studio component is being created, modified, or deleted, its state will equal "CREATE_IN_PROGRESS", "UPDATE_IN_PROGRESS", or "DELETE_IN_PROGRESS"

These are called 'transition states'.

No modifications may be made to the studio component while it is in a transition state.

If creation of the resource fails, the state will change to `CREATE_FAILED`. The resource StatusCode and StatusMessage will provide more information of why creation failed. The resource in this state will automatically be deleted from your account after a period of time.

If updating the resource fails, the state will change to `UPDATE_FAILED`. The resource StatusCode and StatusMessage will provide more information of why the update failed. The resource will be returned to the state it was in when the update request was invoked.

If deleting the resource fails, the state will change to `DELETE_FAILED`. The resource StatusCode and StatusMessage will provide more information of why the update failed. The resource will be returned to the state it was in when the update request was invoked. After the resource is deleted successfully, it will change to the "DELETED" state. The resource will no longer count against service quotas and cannot be used or acted upon any futher. It will be removed from your account after a period of time.

StudioComponentStatusCode

The current status of the studio component resource.

When the resource is in the 'READY' state, the status code signals what the last mutation made to the resource was.

When the resource is in a CREATE_FAILED/UPDATE_FAILED/DELETE_FAILED state, the status code signals what went wrong and why the mutation failed.

StudioComponentSubtype

StudioComponentType

StudioEncryptionConfigurationKeyType

The type of KMS key that is used to encrypt studio data.

StudioPersona

StudioState

StudioStatusCode

The status code.