Module aws_sdk_nimble::operation

Structs

AcceptEulas

Accept EULAs.

CreateLaunchProfile

Create a launch profile.

CreateStreamingImage

Creates a streaming image resource in a studio.

CreateStreamingSession

Creates a streaming session in a studio.

After invoking this operation, you must poll GetStreamingSession until the streaming session is in state READY.

CreateStreamingSessionStream

Creates a streaming session stream for a streaming session.

After invoking this API, invoke GetStreamingSessionStream with the returned streamId to poll the resource until it is in state READY.

CreateStudio

Create a new Studio.

When creating a Studio, two IAM roles must be provided: the admin role and the user Role. These roles are assumed by your users when they log in to the Nimble Studio portal.

The user role must have the AmazonNimbleStudio-StudioUser managed policy attached for the portal to function properly.

The Admin Role must have the AmazonNimbleStudio-StudioAdmin managed policy attached for the portal to function properly.

You may optionally specify a KMS key in the StudioEncryptionConfiguration.

In Nimble Studio, resource names, descriptions, initialization scripts, and other data you provide are always encrypted at rest using an KMS key. By default, this key is owned by Amazon Web Services and managed on your behalf. You may provide your own KMS key when calling CreateStudio to encrypt this data using a key you own and manage.

When providing an KMS key during studio creation, Nimble Studio creates KMS grants in your account to provide your studio user and admin roles access to these KMS keys.

If you delete this grant, the studio will no longer be accessible to your portal users.

If you delete the studio KMS key, your studio will no longer be accessible.

CreateStudioComponent

Creates a studio component resource.

DeleteLaunchProfile

Permanently delete a launch profile.

DeleteLaunchProfileMember

Delete a user from launch profile membership.

DeleteStreamingImage

Delete streaming image.

DeleteStreamingSession

Deletes streaming session resource.

After invoking this operation, use GetStreamingSession to poll the resource until it transitions to a DELETED state.

A streaming session will count against your streaming session quota until it is marked DELETED.

DeleteStudio

Delete a studio resource.

DeleteStudioComponent

Deletes a studio component resource.

DeleteStudioMember

Delete a user from studio membership.

GetEula

Get Eula.

GetLaunchProfile

Get a launch profile.

GetLaunchProfileDetails

Launch profile details include the launch profile resource and summary information of resources that are used by, or available to, the launch profile. This includes the name and description of all studio components used by the launch profiles, and the name and description of streaming images that can be used with this launch profile.

GetLaunchProfileInitialization

Get a launch profile initialization.

GetLaunchProfileMember

Get a user persona in launch profile membership.

GetStreamingImage

Get streaming image.

GetStreamingSession

Gets StreamingSession resource.

Invoke this operation to poll for a streaming session state while creating or deleting a session.

GetStreamingSessionStream

Gets a StreamingSessionStream for a streaming session.

Invoke this operation to poll the resource after invoking CreateStreamingSessionStream.

After the StreamingSessionStream changes to the state READY, the url property will contain a stream to be used with the DCV streaming client.

GetStudio

Get a Studio resource.

GetStudioComponent

Gets a studio component resource.

GetStudioMember

Get a user's membership in a studio.

ListEulaAcceptances

List Eula Acceptances.

ListEulas

List Eulas.

ListLaunchProfileMembers

Get all users in a given launch profile membership.

ListLaunchProfiles

List all the launch profiles a studio.

ListStreamingImages

List the streaming image resources available to this studio.

This list will contain both images provided by Amazon Web Services, as well as streaming images that you have created in your studio.

ListStreamingSessions

Lists the streaming image resources in a studio.

ListStudioComponents

Lists the StudioComponents in a studio.

ListStudioMembers

Get all users in a given studio membership.

ListStudios

List studios in your Amazon Web Services account in the requested Amazon Web Services Region.

ListTagsForResource

Gets the tags for a resource, given its Amazon Resource Names (ARN).

This operation supports ARNs for all resource types in Nimble Studio that support tags, including studio, studio component, launch profile, streaming image, and streaming session. All resources that can be tagged will contain an ARN property, so you do not have to create this ARN yourself.

PutLaunchProfileMembers

Add/update users with given persona to launch profile membership.

PutStudioMembers

Add/update users with given persona to studio membership.

StartStudioSSOConfigurationRepair

Repairs the SSO configuration for a given studio.

If the studio has a valid Amazon Web Services SSO configuration currently associated with it, this operation will fail with a validation error.

If the studio does not have a valid Amazon Web Services SSO configuration currently associated with it, then a new Amazon Web Services SSO application is created for the studio and the studio is changed to the READY state.

After the Amazon Web Services SSO application is repaired, you must use the Amazon Nimble Studio console to add administrators and users to your studio.

TagResource

Creates tags for a resource, given its ARN.

UntagResource

Deletes the tags for a resource.

UpdateLaunchProfile

Update a launch profile.

UpdateLaunchProfileMember

Update a user persona in launch profile membership.

UpdateStreamingImage

Update streaming image.

UpdateStudio

Update a Studio resource.

Currently, this operation only supports updating the displayName of your studio.

UpdateStudioComponent

Updates a studio component resource.